ListFragment 상세내역 기능 2차
